Updates to compile with MinGW 7.2  + small fixes (#13)
* Added missing headers to compile with MinGW 7.2. Added support for Glad instead of GLEW with USE_GLAD.

* [BlockFurnace] Fixed smelting timer. Bit shifting u16(32) by <<16(32) is undefined behaviour.

* #undef M_PI and redefine to ensure portability

+++ source/world/BlockFurnace.cpp
@@ -86,6 +86,6 @@ void BlockFurnace::onTick(const glm::ivec3 &blockPosition, Player &, Chunk &, Wo
 		data->inventory.setStack(1, 0, recipe->result().item().id(), outputStack.amount() + recipe->result().amount());
 	}
 
-	data->data = ticksRemaining | (currentBurnTime << 16) | ((u32)itemProgress << 32);
+	data->data = ticksRemaining | ((u32)currentBurnTime << 16) | ((u64)itemProgress << 32);
 }
